Finished working on welcomingPanel

+++ Muniffic-ScriptCore/Source/eg/Scene/Components.cs
@@ -647,13 +647,13 @@ namespace eg
 
         public bool isSensor
         {
-            get => InternalCalls.BoxCollider2DComponent_IsSensor(Entity.ID);
-            set => InternalCalls.BoxCollider2DComponent_SetSensor(Entity.ID, ref value);
+            get => InternalCalls.BoxCollider2DComponent_IsSensor((ulong)Entity.ID);
+            set => InternalCalls.BoxCollider2DComponent_SetSensor((ulong)Entity.ID, ref value);
         }
 
         public bool CollidesWith(Entity other)
         {
-            return InternalCalls.BoxCollider2DComponent_CollidesWith(Entity.ID, other.ID);
+            return InternalCalls.BoxCollider2DComponent_CollidesWith((ulong)Entity.ID, (ulong)other.ID);
         }
 
         public bool CollidesWithBox(Entity other)
@@ -709,7 +709,7 @@ namespace eg
 
         public bool CollidesWith(long entityID)
         {
-            return InternalCalls.BoxCollider2DComponent_CollidesWith(Entity.ID, entityID);
+            return InternalCalls.BoxCollider2DComponent_CollidesWith((ulong)Entity.ID, entityID);
         }
 
         bool CollidesWithEntitiesSide(Entity entity, Side side)
@@ -785,13 +785,13 @@ namespace eg
         }
         public bool isSensor
         {
-            get => InternalCalls.CircleCollider2DComponent_IsSensor(Entity.ID);
-            set => InternalCalls.CircleCollider2DComponent_SetSensor(Entity.ID, ref value);
+            get => InternalCalls.CircleCollider2DComponent_IsSensor((ulong)Entity.ID);
+            set => InternalCalls.CircleCollider2DComponent_SetSensor((ulong)Entity.ID, ref value);
         }
 
         public bool CollidesWith(Entity other)
         {
-            return InternalCalls.CircleCollider2DComponent_CollidesWith(Entity.ID, other.ID);
+            return InternalCalls.CircleCollider2DComponent_CollidesWith((ulong)Entity.ID, (ulong)other.ID);
         }
 
         public bool CollidesWithCircle(Entity other)
@@ -822,7 +822,7 @@ namespace eg
         public bool CollidesWithEntity(long entityID)
         {
 
-            return InternalCalls.CircleCollider2DComponent_CollidesWith(Entity.ID, entityID);
+            return InternalCalls.CircleCollider2DComponent_CollidesWith((ulong)Entity.ID, (ulong)entityID);
         }
 
         public bool CollidesWithEntity(string entityName)
